added vertices to geometry data

+++ RvtVa3c/Va3cScene.cs
@@ -81,7 +81,7 @@ namespace RvtVa3c
     [DataContract]
     public class Va3cGeometryData
     {
-      [DataMember] public List<int> vertices { get; set; } // millimetres
+      [DataMember] public List<long> vertices { get; set; } // millimetres
       // "morphTargets": []
       [DataMember] public List<double> normals { get; set; }
       // "colors": []
